Added edit profile forms, views and templates

+++ farm_food_project/profiles/forms.py
@@ -12,9 +12,15 @@ class ProducerProfileForm(forms.ModelForm):
         model = ProducerUserProfile
         exclude = ('user',)
 
+
+class EditProducerProfileForm(ProducerProfileForm):
+    class Meta:
+        model = ProducerUserProfile
+        exclude = ('user',)
+
     def save(self, commit=True):
-        db_producer = ProducerUserProfile.objects.get(pk=self.instance.id)
-        if commit:
+        db_producer = ProducerUserProfile.objects.get(pk=self.instance.user_id)
+        if commit and db_producer.profile_image:
             os.remove(join(settings.MEDIA_ROOT, str(db_producer.profile_image)))
         return super().save(commit)
 
@@ -23,3 +29,15 @@ class ConsumerProfileForm(forms.ModelForm):
     class Meta:
         model = ConsumerUserProfile
         exclude = ('user',)
+
+
+class EditConsumerProfileForm(ConsumerProfileForm):
+    class Meta:
+        model = ConsumerUserProfile
+        exclude = ('user',)
+
+    def save(self, commit=True):
+        db_consumer = ConsumerUserProfile.objects.get(pk=self.instance.user_id)
+        if commit and db_consumer.profile_image:
+            os.remove(join(settings.MEDIA_ROOT, str(db_consumer.profile_image)))
+        return super().save(commit)
